+++ basic-user-auth.glopen.js
@@ -2,7 +2,7 @@ export default {
 	merge: [
 		{
 			// required for the models shared across routes
-			dir: './shared',
+			dir: './_shared',
 		},
 		{
 			dir: './routes/basic-user-auth',