+++ definition/single-user/openapi/tags.@.js
@@ -1,0 +1,4 @@
+export const singleUser = {
+	name: 'singleUser',
+	description: 'Routes that handle things for a single user, namely user creation and updating ones own user properties.',
+}