+++ definition/user-api-tokens/openapi/tags.@.js
@@ -1,0 +1,4 @@
+export const userApiTokens = {
+	description: 'Routes that handle managing API tokens that are used by individual users.',
+	name: 'userApiTokens',
+}