+++ definition/user-sessions/openapi/tags.@.js
@@ -1,0 +1,4 @@
+export const userSessions = {
+	name: 'userSessions',
+	description: 'Routes that handle things like user creation, logging in, and session management.',
+}