+++ index.js
@@ -2,7 +2,7 @@ import { dirname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 const __dirname = join(dirname(fileURLToPath(import.meta.url)), 'definition')
-const make = name => ([
+const make = (name, options) => ([
 	{
 		dir: join(__dirname, name,'openapi'),
 		ext: '@',
@@ -10,19 +10,20 @@ const make = name => ([
 	{
 		dir: join(__dirname, name, 'routes'),
 		ext: '@',
+		api: options?.api,
 	},
 ])
 
 export const shared = () => [{ dir: join(__dirname, '_shared', 'openapi'), ext: '@' }]
-export const singleUser = () => make('single-user')
-export const userApiTokens = () => make('user-api-tokens')
-export const userManagement = () => make('user-management')
-export const userSessions = () => make('user-sessions')
+export const singleUser = options => make('single-user', options)
+export const userApiTokens = options => make('user-api-tokens', options)
+export const userManagement = options => make('user-management', options)
+export const userSessions = options => make('user-sessions', options)
 
-export const all = () => ([
+export const all = options => ([
 	...shared(),
-	...singleUser(),
-	...userApiTokens(),
-	...userManagement(),
-	...userSessions(),
+	...singleUser(options),
+	...userApiTokens(options),
+	...userManagement(options),
+	...userSessions(options),
 ])