+++ definition/_shared/openapi/components/responses/error.@.js
@@ -1,0 +1,10 @@
+export default {
+	description: 'A list of one or more errors from the server.',
+	content: {
+		'application/json': {
+			schema: {
+				$ref: '#/components/schemas/errorBody',
+			},
+		},
+	},
+}