+++ definition/_shared/openapi/components/schemas/errorBody.@.js
@@ -1,0 +1,12 @@
+export default {
+	description: 'The response body when the only thing returned is one or more errors.',
+	type: 'object',
+	properties: {
+		error: {
+			type: 'array',
+			items: {
+				$ref: '#/components/schemas/error',
+			},
+		},
+	},
+}