+++ definition/_shared/openapi/components/securitySchemes/cookie.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: `
+Session-based authentication, typically browser based, which requires the user to
+authenticate using the \`POST /sessions\` route, and 2FA if enabled.
+	`,
+	type: 'apiKey',
+	name: 'SESSION',
+	in: 'cookie',
+}