+++ definition/user-api-tokens/openapi/components/parameters/userApiTokenId.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: 'The identifier of the specific API token owned by a single user.',
+	name: 'userApiTokenId',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}