+++ definition/user-management/openapi/components/parameters/userId.@.js
@@ -1,0 +1,9 @@
+export default {
+	description: 'The identifier of the specific user.',
+	name: 'userId',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}