+++ definition/user-sessions/openapi/components/parameters/sessionId.@.js
@@ -1,0 +1,9 @@
+export default {
+	name: 'sessionId',
+	description: 'The identifier of the specific session.',
+	in: 'path',
+	required: true,
+	schema: {
+		type: 'string',
+	},
+}