+++ index.js
@@ -1,0 +1,29 @@
+import { dirname, join } from 'node:path'
+import { fileURLToPath } from 'node:url'
+
+const __dirname = join(dirname(fileURLToPath(import.meta.url)), 'definition')
+const make = (name, options) => ([
+	{
+		dir: join(__dirname, name,'openapi'),
+		ext: '@',
+	},
+	{
+		dir: join(__dirname, name, 'routes'),
+		ext: '@',
+		api: options?.api,
+	},
+])
+
+export const shared = () => [{ dir: join(__dirname, '_shared', 'openapi'), ext: '@' }]
+export const singleUser = options => make('single-user', options)
+export const userApiTokens = options => make('user-api-tokens', options)
+export const userManagement = options => make('user-management', options)
+export const userSessions = options => make('user-sessions', options)
+
+export const all = options => ([
+	...shared(),
+	...singleUser(options),
+	...userApiTokens(options),
+	...userManagement(options),
+	...userSessions(options),
+])