+++ routes/basic-user-auth/tags.@.js
@@ -1,4 +1,0 @@
-export const basicUserAuth = {
-	name: 'basicUserAuth',
-	description: 'Routes that handle things like user creation, logging in, and session management.',
-}