+++ CHANGELOG.md
@@ -22,6 +22,14 @@ Change categories are:
 ### Removed
 ### Security
 
+## [0.0.2] - 2021-11-04
+### Added
+- Security blocks for each endpoint.
+- Routes to handle management of API tokens.
+- Routes to handle managing users.
+### Changed
+- BREAKING CHANGE: The paths changed, so that you can more easily re-export routes, making it possible to pick single Operation Objects instead of all.
+
 ## [0.0.1] - 2021-10-27
 ### Added
 - The main basic user auth related routes that I use all over the place.